constant: move constant* to constant.h
